fix: schema.sql type consistency, rename credits label for clarity
Agent-Logs-Url: https://github.com/ujjww4ll/SmartCanteenSystem/sessions/a2340ccc-15bd-4442-85d4-741cd814f483

+++ database/schema.sql
@@ -2,7 +2,7 @@ CREATE TABLE orders (
     order_id INT,
     status TEXT,
     expected_time INT,
-    actual_time INT,
+    actual_time REAL,
     final_price INT,
     preparing_time REAL
 );
